Struct sfcgal::SFCGeometry

#[repr(C)]
pub struct SFCGeometry { /* private fields */ }

Object representing a SFCGAL Geometry.

Most of the operations allowed by SFCGAL C API are wrapped, except those modifying the geometry in-place (such as adding a new point to a linestring for example) and those retrieving a specific part of a geometry (such as getting the 2nd interior ring of some polygon as a linestring). However, this can easily be done by yourself by converting them from/to coordinates with the new_from_coordinates and to_coordinates methods.

(C API reference)

Implementations

impl SFCGeometry

pub fn new(wkt: &str) -> Result<SFCGeometry>

Create a geometry by parsing a WKT string.

pub fn new_from_coordinates<T>(coords: &CoordSeq<T>) -> Result<SFCGeometry>where
T: ToSFCGALGeom + CoordType,

pub fn to_wkt(&self) -> Result<String>

Returns a WKT representation of the given SFCGeometry using CGAL exact integer fractions as coordinate values. (C API reference)

pub fn to_wkt_decim(&self, nb_decim: i32) -> Result<String>

Returns a WKT representation of the given SFCGeometry using floating point coordinate values with the desired number of decimals. (C API reference)

pub fn is_empty(&self) -> Result<bool>

Test if the given SFCGeometry is empty or not.

pub fn is_valid(&self) -> Result<bool>

Test if the given SFCGeometry is valid or not.

pub fn is_measured(&self) -> Result<bool>

Test if the given SFCGeometry is measured (has an ‘m’ coordinates)

pub fn validity_detail(&self) -> Result<Option<String>>

Returns reason for the invalidity or None in case of validity.

pub fn is_planar(&self) -> Result<bool>

Test if the given SFCGeometry is planar or not.

pub fn is_3d(&self) -> Result<bool>

Test if the given SFCGeometry is a 3d geometry or not.

pub fn _type(&self) -> Result<GeomType>

Returns the SFCGAL type of the given SFCGeometry.

pub fn distance(&self, other: &SFCGeometry) -> Result<f64>

Computes the distance to an other SFCGeometry.

pub fn distance_3d(&self, other: &SFCGeometry) -> Result<f64>

Computes the 3d distance to an other SFCGeometry.

pub fn area(&self) -> Result<f64>

Computes the area of the given SFCGeometry.

pub fn area_3d(&self) -> Result<f64>

Computes the 3d area of the given SFCGeometry.

pub fn volume(&self) -> Result<f64>

Computes the volume of the given SFCGeometry (must be a volume).

pub fn orientation(&self) -> Result<Orientation>

Computes the orientation of the given SFCGeometry (must be a Polygon)

pub fn intersects(&self, other: &SFCGeometry) -> Result<bool>

Test the intersection with an other SFCGeometry.

pub fn intersects_3d(&self, other: &SFCGeometry) -> Result<bool>

Test the 3d intersection with an other SFCGeometry.

pub fn intersection(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the intersection of the given SFCGeometry to an other SFCGeometry.

pub fn intersection_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d intersection of the given SFCGeometry to an other SFCGeometry.

pub fn covers(&self, other: &SFCGeometry) -> Result<bool>

pub fn covers_3d(&self, other: &SFCGeometry) -> Result<bool>

pub fn difference(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the difference of the given SFCGeometry to an other SFCGeometry.

pub fn difference_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d difference of the given SFCGeometry to an other SFCGeometry.

pub fn union(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the union of the given SFCGeometry to an other SFCGeometry.

pub fn union_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d union of the given SFCGeometry to an other SFCGeometry.

pub fn minkowski_sum(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the minkowski sum of the given SFCGeometry and an other SFCGEOMETRY. (C API reference)

pub fn straight_skeleton(&self) -> Result<SFCGeometry>

Returns the straight skeleton of the given SFCGeometry. (C API reference)

pub fn straight_skeleton_distance_in_m(&self) -> Result<SFCGeometry>

Returns the straight skeleton of the given SFCGeometry with the distance to the border as M coordinate. (C API reference)

pub fn approximate_medial_axis(&self) -> Result<SFCGeometry>

Returns the approximate medial axis for the given SFCGeometry Polygon. (C API reference)

pub fn offset_polygon(&self, radius: f64) -> Result<SFCGeometry>

Returns the offset polygon of the given SFCGeometry. (C API reference)

pub fn extrude(&self, ex: f64, ey: f64, ez: f64) -> Result<SFCGeometry>

Returns the extrusion of the given SFCGeometry (not supported on Solid and Multisolid). (C API reference

pub fn tesselate(&self) -> Result<SFCGeometry>

Returns a tesselation of the given SFCGeometry. (C API reference)

pub fn triangulate_2dz(&self) -> Result<SFCGeometry>

Returns a triangulation of the given SFCGeometry. (C API reference)

pub fn convexhull(&self) -> Result<SFCGeometry>

Returns the convex hull of the given SFCGeometry. (C API reference)

pub fn convexhull_3d(&self) -> Result<SFCGeometry>

Returns the 3d convex hull of the given SFCGeometry. (C API reference)

pub fn line_substring(&self, start: f64, end: f64) -> Result<SFCGeometry>

Returns the substring of the given SFCGeometry LineString between fractional distances. (C API reference)

pub fn alpha_shapes(&self, alpha: f64, allow_holes: bool) -> Result<SFCGeometry>

Returns the alpha shape of the given SFCGeometry Point set. (C API reference)

pub fn optimal_alpha_shapes(
&self,
allow_holes: bool,
nb_components: usize
) -> Result<SFCGeometry>

Return the optimal alpha shape of the given SFCGeometry Point set.

pub fn create_collection(geoms: &mut [SFCGeometry]) -> Result<SFCGeometry>

Create a SFCGeometry collection type (MultiPoint, MultiLineString, MultiPolygon, MultiSolid or GeometryCollection) given a mutable slice of SFCGeometry’s (this is a destructive operation)

use sfcgal::SFCGeometry;
let a = SFCGeometry::new("POINT(1.0 1.0)").unwrap();
let b = SFCGeometry::new("POINT(2.0 2.0)").unwrap();
let g = SFCGeometry::create_collection(&mut[a, b]).unwrap();
assert_eq!(
    g.to_wkt_decim(1).unwrap(),
    "MULTIPOINT((1.0 1.0),(2.0 2.0))",
);

pub fn get_collection_members(self) -> Result<Vec<SFCGeometry>>

Get the members of a SFCGeometry. Returns Err if the SFCGeometry if not a collection (i.e. if it’s type is not in { MultiPoint, MultiLineString, MultiPolygon, MultiSolid, GeometryCollection }). The original geometry stay untouched.

use sfcgal::SFCGeometry;
let g = SFCGeometry::new("MULTIPOINT((1.0 1.0),(2.0 2.0))").unwrap();
let members = g.get_collection_members().unwrap();
assert_eq!(
    members[0].to_wkt_decim(1).unwrap(),
    "POINT(1.0 1.0)",
);
assert_eq!(
    members[1].to_wkt_decim(1).unwrap(),
    "POINT(2.0 2.0)",
);