sfcgal

Struct SFCGeometry

#[repr(C)]
pub struct SFCGeometry { /* private fields */ }

Object representing a SFCGAL Geometry.

Implementations

impl SFCGeometry

pub fn new(wkt: &str) -> Result<SFCGeometry>

Create a geometry by parsing a WKT string.

pub fn new_from_coordinates<T>(coords: &CoordSeq<T>) -> Result<SFCGeometry>
where T: ToSFCGALGeom + CoordType,

pub fn to_wkt(&self) -> Result<String>

Returns a WKT representation of the given SFCGeometry using CGAL exact integer fractions as coordinate values.

pub fn to_wkt_decim(&self, nb_decim: i32) -> Result<String>

Returns a WKT representation of the given SFCGeometry using floating point coordinate values with the desired number of decimals.

pub fn to_wkb_in_memory(&self) -> Result<Vec<u8>>

Creates a Wkb string of the given geometry. In memory version.

pub unsafe fn to_ewkt_in_memory( prepared: *const sfcgal_prepared_geometry_t, num_decimals: i32, ) -> Result<Vec<u8>>

Safety

Returns an EWKT representation of the given PreparedGeometry

pub fn to_hexwkb_in_memory(&self) -> Result<Vec<u8>>

Creates a Hexwkb string of the given geometry. In memory version.

pub fn io_read_binary_prepared( data: &[u8], ) -> Result<*mut sfcgal_prepared_geometry_t>

Read the binary prepared geometry

pub fn io_read_wkb(data: &[u8]) -> Result<*mut sfcgal_prepared_geometry_t>

Read the binary prepared geometry

pub fn io_read_ewkt(data: &[u8]) -> Result<*mut sfcgal_prepared_geometry_t>

Read the binary prepared geometry

pub unsafe fn io_write_binary_prepared( prepared_geometry: *mut sfcgal_prepared_geometry_t, ) -> Result<Vec<u8>>

Safety

Read the binary prepared geometry in memory

pub fn to_obj_file(&self, filename: &str) -> Result<()>

Creates a OBJ file of the given geometry

pub fn to_obj_in_memory(&self) -> Result<Vec<u8>>

Creates a OBJ string of the given geometry. In memory version.

pub fn to_vtk_file(&self, filename: &str) -> Result<()>

Creates a VTK file of the given geometry

pub fn to_vtk_in_memory(&self) -> Result<Vec<u8>>

Creates a VTK string of the given geometry. In memory version.

pub fn is_empty(&self) -> Result<bool>

Test if the given SFCGeometry is empty or not.

pub fn is_valid(&self) -> Result<bool>

Test if the given SFCGeometry is valid or not.

pub fn is_measured(&self) -> Result<bool>

Test if the given SFCGeometry is measured (has an ‘m’ coordinates)

pub fn is_planar(&self) -> Result<bool>

Test if the given SFCGeometry is planar or not.

pub fn is_3d(&self) -> Result<bool>

Test if the given SFCGeometry is a 3d geometry or not.

pub fn validity_detail(&self) -> Result<Option<String>>

Returns reason for the invalidity or None in case of validity.

pub fn _type(&self) -> Result<GeomType>

Returns the SFCGAL type of the given SFCGeometry.

pub fn distance(&self, other: &SFCGeometry) -> Result<f64>

Computes the distance to an other SFCGeometry.

pub fn distance_3d(&self, other: &SFCGeometry) -> Result<f64>

Computes the 3d distance to an other SFCGeometry.

pub fn area(&self) -> Result<f64>

Computes the area of the given SFCGeometry.

pub fn area_3d(&self) -> Result<f64>

Computes the 3d area of the given SFCGeometry.

pub fn volume(&self) -> Result<f64>

Computes the volume of the given SFCGeometry (must be a volume).

pub fn orientation(&self) -> Result<Orientation>

Computes the orientation of the given SFCGeometry (must be a Polygon)

pub fn intersects(&self, other: &SFCGeometry) -> Result<bool>

Test the intersection with an other SFCGeometry.

pub fn intersects_3d(&self, other: &SFCGeometry) -> Result<bool>

Test the 3d intersection with an other SFCGeometry.

pub fn intersection(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the intersection of the given SFCGeometry to an other SFCGeometry.

pub fn intersection_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d intersection of the given SFCGeometry to an other SFCGeometry.

pub fn covers(&self, other: &SFCGeometry) -> Result<bool>

Test if the geometry covers an other SFCGeometry.

pub fn covers_3d(&self, other: &SFCGeometry) -> Result<bool>

Test if the geometry covers an other SFCGeometry in 3D.

pub fn difference(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the difference of the given SFCGeometry to an other SFCGeometry.

pub fn difference_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d difference of the given SFCGeometry to an other SFCGeometry.

pub fn union(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the union of the given SFCGeometry to an other SFCGeometry.

pub fn union_3d(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the 3d union of the given SFCGeometry to an other SFCGeometry.

pub fn minkowski_sum(&self, other: &SFCGeometry) -> Result<SFCGeometry>

Returns the minkowski sum of the given SFCGeometry and an other SFCGEOMETRY.

pub fn straight_skeleton(&self) -> Result<SFCGeometry>

Returns the straight skeleton of the given SFCGeometry.

pub fn straight_skeleton_distance_in_m(&self) -> Result<SFCGeometry>

Returns the straight skeleton of the given SFCGeometry with the distance to the border as M coordinate.

pub fn extrude_straight_skeleton(&self, height: f64) -> Result<SFCGeometry>

Returns the extrude straight skeleton of the given Polygon.

pub fn extrude_polygon_straight_skeleton( &self, building_height: f64, roof_height: f64, ) -> Result<SFCGeometry>

Returns the union of the polygon z-extrusion (with respect to building_height) and the extrude straight skeleton (with respect to roof_height) of the given Polygon.

pub fn approximate_medial_axis(&self) -> Result<SFCGeometry>

Returns the approximate medial axis for the given SFCGeometry Polygon.

pub fn offset_polygon(&self, radius: f64) -> Result<SFCGeometry>

Returns the offset polygon of the given SFCGeometry.

pub fn extrude(&self, ex: f64, ey: f64, ez: f64) -> Result<SFCGeometry>

Returns the extrusion of the given SFCGeometry (not supported on Solid and Multisolid).

pub fn tesselate(&self) -> Result<SFCGeometry>

Returns a tesselation of the given SFCGeometry.

pub fn triangulate_2dz(&self) -> Result<SFCGeometry>

Returns a triangulation of the given SFCGeometry.

pub fn convexhull(&self) -> Result<SFCGeometry>

Returns the convex hull of the given SFCGeometry.

pub fn convexhull_3d(&self) -> Result<SFCGeometry>

Returns the 3d convex hull of the given SFCGeometry.

pub fn line_substring(&self, start: f64, end: f64) -> Result<SFCGeometry>

Returns the substring of the given SFCGeometry LineString between fractional distances.

pub fn alpha_shapes(&self, alpha: f64, allow_holes: bool) -> Result<SFCGeometry>

Returns the alpha shape of the given SFCGeometry Point set.

pub fn optimal_alpha_shapes( &self, allow_holes: bool, nb_components: usize, ) -> Result<SFCGeometry>

Return the optimal alpha shape of the given SFCGeometry Point set.

pub fn create_collection(geoms: &mut [SFCGeometry]) -> Result<SFCGeometry>

Create a SFCGeometry collection type (MultiPoint, MultiLineString, MultiPolygon, MultiSolid or GeometryCollection) given a mutable slice of SFCGeometry’s (this is a destructive operation).

use sfcgal::SFCGeometry;
let a = SFCGeometry::new("POINT (1.0 1.0)").unwrap();
let b = SFCGeometry::new("POINT (2.0 2.0)").unwrap();
let g = SFCGeometry::create_collection(&mut[a, b]).unwrap();
assert_eq!(
    g.to_wkt_decim(1).unwrap(),
    "MULTIPOINT ((1.0 1.0),(2.0 2.0))",
);

pub fn get_collection_members(self) -> Result<Vec<SFCGeometry>>

Get the members of a SFCGeometry. Returns Err if the SFCGeometry if not a collection (i.e. if it’s type is not in { MultiPoint, MultiLineString, MultiPolygon, MultiSolid, GeometryCollection }). The original geometry stay untouched.

use sfcgal::SFCGeometry;
let g = SFCGeometry::new("MULTIPOINT ((1.0 1.0),(2.0
2.0))").unwrap(); let members =
g.get_collection_members().unwrap(); assert_eq!(
    members[0].to_wkt_decim(1).unwrap(),
    "POINT (1.0 1.0)",
);
assert_eq!(
    members[1].to_wkt_decim(1).unwrap(),
    "POINT (2.0 2.0)",
);

pub fn point_create() -> Result<SFCGeometry>

Creates an empty point

pub fn point_create_from_xy(x: f64, y: f64) -> Result<SFCGeometry>

Creates a point from two X and Y coordinates

pub fn point_create_from_xym(x: f64, y: f64, m: f64) -> Result<SFCGeometry>

Creates a point from three X, Y and M coordinates

pub fn point_create_from_xyz(x: f64, y: f64, z: f64) -> Result<SFCGeometry>

Creates a point from three X, Y and Z coordinates

pub fn point_create_from_xyzm( &self, x: f64, y: f64, z: f64, m: f64, ) -> Result<SFCGeometry>

Create a point from x, y, z, m components.

pub fn point_x(&self) -> Result<f64>

Returns the X coordinate of the given Point

pub fn point_y(&self) -> Result<f64>

Returns the Y coordinate of the given Point

pub fn point_z(&self) -> Result<f64>

Returns the Z coordinate of the given Point

pub fn point_m(&self) -> Result<f64>

Returns the M coordinate of the given Point

pub fn triangle_set_vertex( &self, index: i32, vertex: &SFCGeometry, ) -> Result<()>

Sets one vertex of a Triangle

pub fn triangle_set_vertex_from_xy( &self, index: i32, x: f64, y: f64, ) -> Result<()>

Sets one vertex of a Triangle from two coordinates

pub fn triangle_set_vertex_from_xyz( &self, index: i32, x: f64, y: f64, z: f64, ) -> Result<()>

Sets one vertex of a Triangle from three coordinates

pub fn straight_skeleton_partition( &self, auto_orientation: bool, ) -> Result<SFCGeometry>

Returns the straight skeleton partition for the given Polygon

pub fn visibility_point(&self, point: &SFCGeometry) -> Result<SFCGeometry>

Returns the visibility polygon of a Point inside a Polygon

pub fn rotate(&self, angle: f64) -> Result<SFCGeometry>

Rotates a geometry around the origin (0,0,0) by a given angle

pub fn rotate_x(&self, angle: f64) -> Result<SFCGeometry>

Rotates a geometry around the X axis by a given angle

pub fn rotate_y(&self, angle: f64) -> Result<SFCGeometry>

Rotates a geometry around the Y axis by a given angle

pub fn rotate_z(&self, angle: f64) -> Result<SFCGeometry>

Rotates a geometry around the Z axis by a given angle

pub fn rotate_2d( &self, angle: f64, origin_x: f64, origin_y: f64, ) -> Result<SFCGeometry>

Rotates a geometry around a specified point by a given angle

pub fn rotate_3d( &self, angle: f64, axis_x_angle: f64, axis_y_angle: f64, axis_z_angle: f64, ) -> Result<SFCGeometry>

Rotates a 3D geometry around a specified axis by a given angle

pub fn rotate_3d_around_center( &self, angle: f64, axis_x_angle: f64, axis_y_angle: f64, axis_z_angle: f64, center_x: f64, center_y: f64, center_z: f64, ) -> Result<SFCGeometry>

Rotates a 3D geometry around a specified axis and center point by a given

pub fn force_rhr(&self) -> Result<SFCGeometry>

Force a Right Handed Rule on the given Geometry

pub fn force_lhr(&self) -> Result<SFCGeometry>

Force a Left Handed Rule on the given Geometry

pub fn scale(&self, scale: f64) -> Result<SFCGeometry>

Scale a geometry by a given factor

pub fn scale_3d( &self, scale_x: f64, scale_y: f64, scale_z: f64, ) -> Result<SFCGeometry>

Scale a geometry by different factors for each dimension

pub fn scale_3d_around_center( &self, factor_x: f64, factor_y: f64, factor_z: f64, center_x: f64, center_y: f64, center_z: f64, ) -> Result<SFCGeometry>

Scale a geometry by different factors for each dimension around a center

pub fn round(&self, value: i32) -> Result<SFCGeometry>

Round coordinates of the given Geometry

pub fn buffer3d( &self, radius: f64, segments: i32, buffer_type: BufferType, ) -> Result<SFCGeometry>

Computes a 3D buffer around a geometry

pub fn has_validity_flag(&self) -> i32

Gets the validity flag of the geometry

pub fn visibility_segment( &self, point_a: &SFCGeometry, point_b: &SFCGeometry, ) -> Result<SFCGeometry>

Build the visibility polygon of the segment [pointA ; pointB] on a Polygon

pub fn make_solid(&self) -> Result<SFCGeometry>

Convert a PolyhedralSurface to a Solid

pub fn translate_2d(&self, dx: f64, dy: f64) -> Result<SFCGeometry>

Translate a geometry by a 2D vector

pub fn translate_3d( &self, translation_x: f64, translation_y: f64, translation_z: f64, ) -> Result<SFCGeometry>

Translate a geometry by a 3D vector

pub fn force_valid(&self, validity: i32)

Sets the validity flag of the geometry.

pub fn set_geometry_validation(enabled: i32)

Set the geometry validation mode

pub fn triangle_create() -> Result<SFCGeometry>

Creates an empty Triangle

pub fn triangle_vertex(&self, index: i32) -> Result<SFCGeometry>

Returns one the Triangle’s vertex as a Point

pub fn triangle_create_from_points( point_a: &SFCGeometry, point_b: &SFCGeometry, point_c: &SFCGeometry, ) -> Result<SFCGeometry>

Creates a Triangle from three given Point

pub fn triangulated_surface_create() -> Result<SFCGeometry>

Creates an empty TriangulatedSurface

pub fn triangulated_surface_triangle_n( &self, index: usize, ) -> Result<SFCGeometry>

Returns the ith Triangle of a given TriangulatedSurface

pub fn triangulated_surface_add_triangle( &self, triangle: &SFCGeometry, ) -> Result<()>

Adds a Triangle to a given TriangulatedSurface

pub fn triangulated_surface_num_triangles(&self) -> Result<usize>

Returns the number of triangles of a given TriangulatedSurface

pub fn optimal_convex_partition_2(&self) -> Result<SFCGeometry>

Returns the optimal convex partition of a geometry (polygon without hole)

pub fn approx_convex_partition_2(&self) -> Result<SFCGeometry>

Returns the approximal convex partition of a geometry (polygon without hole)

pub fn greene_approx_convex_partition_2(&self) -> Result<SFCGeometry>

Returns the greene approximal convex partition of a geometry (polygon without hole)

pub fn y_monotone_partition_2(&self) -> Result<SFCGeometry>

Returns the y monotone partition of a geometry (polygon without hole)

pub fn linestring_num_points(&self) -> Result<usize>

Returns the number of points of the given LineString

pub fn linestring_add_point(&self, point: &SFCGeometry) -> Result<()>

Adds a point to a LineString

pub fn linestring_point_n(&self, index: usize) -> Result<SFCGeometry>

Returns the ith point of a given LineString

pub fn linestring_create() -> Result<SFCGeometry>

Creates an empty LineString

pub fn polyhedral_surface_create() -> Result<SFCGeometry>

Creates an empty PolyhedralSurface

pub fn polyhedral_surface_num_polygons(&self) -> Result<usize>

Returns the number of polygons of a given PolyhedralSurface

pub fn polyhedral_surface_add_polygon(&self, other: &SFCGeometry) -> Result<()>

Adds a Polygon to a given PolyhedralSurface

pub fn polyhedral_surface_polygon_n(&self, index: usize) -> Result<SFCGeometry>

Returns the ith polygon of a given PolyhedralSurface

pub fn set_error_handlers( warning_handler: sfcgal_error_handler_t, error_handler: sfcgal_error_handler_t, )

Sets the error handlers. These callbacks are called on warning or error

pub fn set_alloc_handlers( alloc_handler: sfcgal_alloc_handler_t, free_handler: sfcgal_free_handler_t, )

Sets the error handlers. These callbacks are called on warning or error

pub fn solid_create() -> Result<SFCGeometry>

Creates an empty Solid

pub fn solid_num_shells(&self) -> Result<usize>

Returns the number of shells of a given Solid

pub fn solid_create_from_exterior_shell( exterior_shell: &SFCGeometry, ) -> Result<SFCGeometry>

Creates a Solid from an exterior shell

pub fn solid_shell_n(&self, index: usize) -> Result<SFCGeometry>

Returns the ith shell of a given Solid

pub fn solid_add_interior_shell(&self, shell: &SFCGeometry) -> Result<()>

Adds a shell to a given Solid

pub fn polygon_create() -> Result<SFCGeometry>

Creates an empty Polygon

pub fn polygon_num_interior_rings(&self) -> Result<usize>

Returns the number of interior rings of a given Polygon

pub fn polygon_exterior_ring(&self) -> Result<SFCGeometry>

Returns the exterior ring of a given Polygon

pub fn polygon_create_from_exterior_ring(&self) -> Result<SFCGeometry>

Creates an empty Polygon from an extrior ring

pub fn polygon_add_interior_ring(&self, ring: &SFCGeometry) -> Result<()>

Adds an interior ring to a given Polygon

pub fn polygon_interior_ring_n(&self, index: usize) -> Result<SFCGeometry>

Returns the ith interior ring of a given Polygon

pub fn multi_solid_create() -> Result<SFCGeometry>

Creates an empty MultiSolid

pub unsafe fn prepared_geometry_set_srid( prepared_geometry: *mut sfcgal_prepared_geometry_t, srid: srid_t, )

Safety

Sets SRID associated with a given PreparedGeometry

pub unsafe fn prepared_geometry_delete( prepared_geometry: *mut sfcgal_prepared_geometry_t, )

Safety

Deletes a given PreparedGeometry

pub fn prepared_geometry_create_from_geometry( &self, srid: srid_t, ) -> Result<*mut sfcgal_prepared_geometry_t>

Safety

Creates a PreparedGeometry from a Geometry and an SRID

pub unsafe fn prepared_geometry_set_geometry( &self, prepared: *mut sfcgal_prepared_geometry_t, )

Safety

Sets the Geometry associated with the given PreparedGeometry

pub unsafe fn prepared_geometry_srid( prepared: *mut sfcgal_prepared_geometry_t, ) -> u32

Safety

Returns SRID associated with a given PreparedGeometry

pub fn prepared_geometry_create() -> Result<*mut sfcgal_prepared_geometry_t>

Creates an empty PreparedGeometry

pub unsafe fn prepared_geometry_geometry( prepared_geometry: *mut sfcgal_prepared_geometry_t, ) -> Result<SFCGeometry>

Safety

Returns the Geometry associated with a given PreparedGeometry